Putting up Pim for an easter egg!
Since I'm on a holiday, Pim tried to create an new JIRA project for DataScience. Although I appreciate the effort, he did not follow the newly created JIRA workflows and schemes for DealerDirect.

This causes JIRA to generate all kinds of stuff, we were trying to get rid of....

Next time Pim, please copy one of the newly created projects on project creation.